+++ odfdo/scripts/show.py
@@ -111,14 +111,14 @@ def dump_pictures(document: Document, target: str | Path) -> None:
 
 def spreadsheet_to_stdout(document: Document) -> None:
     body = document.body
-    for table in body.get_tables():
+    for table in body.tables:
         table.rstrip(aggressive=True)  # type: ignore
         print(table.to_csv(None))  # type: ignore
 
 
 def spreadsheet_to_csv(document: Document, output: Path) -> None:
     body = document.body
-    for table in body.get_tables():
+    for table in body.tables:
         name = table.name  # type: ignore
         filename = clean_filename(name) + ".csv"
         print(filename)